RRD-34: Unified interface and enum on one folder
