--- src/interfaces/authServices.ts
@@ -1,4 +1,0 @@
-export interface UserLogin {
-  email: string;
-  password: string;
-}
