+++ src/navigation/mainStack/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import useIsSignedIn from '../../hooks/isSignIn';
 import {appNavOptions, mainStackNavOption} from '../../helpers';
-import {Routes, StackParamList} from '../../interfaces';
+import {Routes, StackParamList} from '../../types/interfaces';
 import TabNavigator from '../tabNavigator';
 import {CreateAccount, Login, SelectPlayers} from '../../screens';
 import LoadingSpinner from '../../components/shared/activityIndicator';
