+++ src/navigation/tabNavigator/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
 import {Match, History, Profile} from '../../screens';
-import {Routes} from '../../interfaces/routes';
+import {Routes} from '../../types/interfaces/routes';
 import {tabIcons, tabIconsFocused} from './constants';
-import {StackParamList} from '../../interfaces';
+import {StackParamList} from '../../types/interfaces';
 
 const Tab = createBottomTabNavigator<StackParamList>();
