+++ src/services/auth.ts
@@ -1,13 +1,12 @@
 import {URL, api} from '../api';
-import {UserLogin} from '../interfaces';
+import {UserLogin} from '../types/interfaces';
 
 export const logInRequest = async (body: UserLogin) => {
   const response = await api.post(URL.LOGIN, body);
-
   return response.data.payload;
 };
 
-export const creatAccountRequest = async (body: UserLogin) => {
-  const response = await api.post(URL.CREATE_ACCCOUNT, body);
+export const createAccountRequest = async (body: UserLogin) => {
+  const response = await api.post(URL.CREATE_ACCOUNT, body);
   return response.data.message;
 };
