+++ src/types/enums/match.ts
@@ -1,0 +1,5 @@
+export enum MatchStatus {
+  toBePlayed = 'To be played',
+  cancelled = 'Cancelled',
+  finished = 'Finished',
+}
