+++ src/types/enums/player.ts
@@ -10,9 +10,3 @@ export enum PlayerStatus {
   starter = 'Starter',
   sustitute = 'Sustitute',
 }
-
-export enum MatchStatus {
-  toBePlayed = 'To be played',
-  cancelled = 'Cancelled',
-  finished = 'Finished',
-}
