+++ src/types/interfaces/authServices.ts
@@ -1,0 +1,8 @@
+import {PlayerPosition} from '../enums';
+
+export interface UserLogin {
+  firstName: string;
+  lastName: string;
+  skill: number;
+  position: PlayerPosition;
+}
